# dist/openapi.yaml
openapi: 3.0.3
info:
  title: Grandma's Recipes API
  version: 1.0.0
  description: |
    This API supports recipe management, user authentication, search, and personal features such as favorites and custom recipes.

servers:
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Auth
    description: Registration, login, and logout endpoints
  - name: Recipes
    description: Recipe retrieval, creation, and search
  - name: User
    description: User-related data (favorites, custom recipes, etc.)
  - name: MealPlan
    description: Meal planning feature
  - name: Preparation
    description: Recipe preparation progress tracking

paths:
  /register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Username already exists

  /login:
    post:
      tags: [Auth]
      summary: Authenticate a user and return token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
        '401':
          description: Invalid credentials

  /logout:
    post:
      tags: [Auth]
      summary: Logout a user and delete their session
      responses:
        '204':
          description: User logged out successfully

  /recipes:
    get:
      tags: [Recipes]
      summary: Get random recipes
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            example: 5
      responses:
        '200':
          description: List of random recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'

    post:
      tags: [Recipes]
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '201':
          description: Recipe created

  /recipes/search:
    get:
      tags: [Recipes]
      summary: Search for recipes
      parameters:
        - in: query
          name: query
          schema:
            type: string
            example: chocolate
        - in: query
          name: cuisine
          schema:
            type: string
            example: Italian
        - in: query
          name: diet
          schema:
            type: string
            example: vegetarian
        - in: query
          name: intolerance
          schema:
            type: string
            example: gluten
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'

  /recipes/{id}:
    get:
      tags: [Recipes]
      summary: Get full recipe information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 42
      responses:
        '200':
          description: Full recipe information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'

  /users/favorites:
    get:
      tags: [User]
      summary: Get user favorite recipes
      responses:
        '200':
          description: List of favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'

    post:
      tags: [User]
      summary: Add recipe to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  example: 101
      responses:
        '201':
          description: Added to favorites

  /users/recipes:
    get:
      tags: [User]
      summary: Get user's created recipes
      responses:
        '200':
          description: User's recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
  
  /users/watched:
    get:
      tags: [User]
      summary: Get user's watched recipes
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            example: 5
          description: Get users latest maximum number of watched recipes to return
      responses:
        '200':
          description: List of watched recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'

    post:
      tags: [User]
      summary: Mark recipe as watched
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipeId:
                  type: integer
                  example: 101
      responses:
        '201':
          description: Recipe marked as watched

components:
  schemas:
    RegisterRequest:
      type: object
      required: [userame, firstname, lastname, country, email, password]
      properties:
        username: { type: string, example: chefAvital }
        firstname: { type: string, example: Avital }
        lastname: { type: string, example: Cohen }
        country: { type: string, example: Israel }
        email: { type: string, format: email, example: avital@example.com }
        password: { type: string, example: S3cretPass! }

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: chefAvital }
        password: { type: string, example: S3cretPass! }

    Recipe:
      type: object
      properties:
        id: { type: integer, example: 101 }
        title: { type: string, example: Grandma's Apple Pie }
        image: { type: string, example: https://example.com/images/apple-pie.jpg }
        duration: { type: integer, example: 90 }
        # likes: { type: integer, example: 452 }
        vegan: { type: boolean, example: false }
        vegetarian: { type: boolean, example: true }
        glutenFree: { type: boolean, example: false }
        viewed: { type: boolean, example: true }
        ingredients:
          type: array
          items:
            type: string
          example:
            - 3 apples, sliced
            - 2 cups flour
            - 1/2 cup sugar
            - 1 tsp cinnamon
        steps:
          type: array
          items:
            type: string
          example:
            - Preheat oven to 180°C
            - Mix sliced apples with sugar and cinnamon
            - Prepare crust and place filling inside
            - Bake for 45 minutes
        servings: { type: integer, example: 6 }
    
    # FamilyRecipe:
    #   type: object
    #   required:
    #     - title
    #     - image
    #     - occasion
    #     - originator_name
    #     - instructions
    #     - ingredients
    #   properties:
    #     title:
    #       type: string
    #       example: Grandma's Apple Pie
    #     image:
    #       type: string
    #       format: uri
    #       example: https://example.com/images/apple-pie.jpg
    #     occasion:
    #       type: string
    #       example: Holiday
    #     originator_name:
    #       type: string
    #       example: Grandma Sarah
    #     instructions:
    #       type: string
    #       example: |
    #         Preheat oven to 180°C
    #         Mix apples with sugar
    #         Bake for 45 minutes
    #     ingredients:
    #       type: string
    #       example: |
    #         3 apples, sliced
    #         2 cups flour
    #         1/2 cup sugar


    # RecipePreview:
    #   type: object
    #   properties:
    #     id:          { type: integer, example: 101 }
    #     title:       { type: string,  example: Pasta Bolognese }
    #     image:       { type: string,  example: https://…/pasta.jpg }
    #     duration:    { type: integer, example: 35 }
    #     vegan:       { type: boolean, example: false }
    #     vegetarian:  { type: boolean, example: true }
    #     glutenFree:  { type: boolean, example: false }
    #     viewed:      { type: boolean, example: true }
    #     favorite:    { type: boolean, example: false }